+++ framework/server/src/main/java/is/codion/framework/server/AbstractRemoteEntityConnection.java
@@ -23,10 +23,14 @@ import is.codion.common.db.exception.DatabaseException;
 import is.codion.common.event.Event;
 import is.codion.common.logging.MethodTrace;
 import is.codion.common.observer.Observer;
+import is.codion.common.property.PropertyValue;
 import is.codion.common.rmi.server.RemoteClient;
 import is.codion.common.user.User;
 import is.codion.framework.db.EntityConnection;
+import is.codion.framework.db.EntityResultIterator;
+import is.codion.framework.db.rmi.RemoteEntityResultIterator;
 import is.codion.framework.domain.Domain;
+import is.codion.framework.domain.entity.Entity;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -39,6 +43,13 @@ import java.rmi.server.RMIClientSocketFactory;
 import java.rmi.server.RMIServerSocketFactory;
 import java.rmi.server.UnicastRemoteObject;
 import java.util.List;
+import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.TimeUnit;
+
+import static is.codion.common.Configuration.longValue;
+import static java.lang.System.currentTimeMillis;
+import static java.util.Collections.newSetFromMap;
 
 /**
  * A base class for remote connections served by a {@link EntityServer}.
@@ -50,6 +61,16 @@ public abstract class AbstractRemoteEntityConnection extends UnicastRemoteObject
 	private static final long serialVersionUID = 1;
 
 	private static final Logger LOG = LoggerFactory.getLogger(AbstractRemoteEntityConnection.class);
+
+	/**
+	 * Specifies the timeout in milliseconds for idle remote iterators.
+	 * Iterators that remain idle for longer than this timeout are automatically closed server-side.
+	 * <ul>
+	 * <li>Value type: Long
+	 * <li>Default value: 300000 (5 minutes)
+	 * </ul>
+	 */
+	static final PropertyValue<Long> ITERATOR_TIMEOUT = longValue("codion.db.remote.iteratorTimeout", TimeUnit.MINUTES.toMillis(5));
 
 	/**
 	 * A Proxy for logging method calls
@@ -65,6 +86,12 @@ public abstract class AbstractRemoteEntityConnection extends UnicastRemoteObject
 	 * An event triggered when this connection is closed
 	 */
 	private final transient Event<AbstractRemoteEntityConnection> closed = Event.event();
+
+	private final Set<DefaultRemoteEntityResultIterator> remoteIterators = newSetFromMap(new ConcurrentHashMap<>());
+
+	private final transient int connectionPort;
+	private final transient RMIClientSocketFactory clientSocketFactory;
+	private final transient RMIServerSocketFactory serverSocketFactory;
 
 	/**
 	 * Instantiates a new AbstractRemoteEntityConnection and exports it on the given port number
@@ -87,6 +114,9 @@ public abstract class AbstractRemoteEntityConnection extends UnicastRemoteObject
 		this.connectionHandler = new LocalConnectionHandler(domain, remoteClient, database);
 		this.connectionProxy = (EntityConnection) Proxy.newProxyInstance(EntityConnection.class.getClassLoader(),
 						new Class[] {EntityConnection.class}, connectionHandler);
+		this.clientSocketFactory = clientSocketFactory;
+		this.serverSocketFactory = serverSocketFactory;
+		this.connectionPort = port;
 	}
 
 	/**
@@ -119,6 +149,7 @@ public abstract class AbstractRemoteEntityConnection extends UnicastRemoteObject
 			catch (NoSuchObjectException e) {
 				LOG.error(e.getMessage(), e);
 			}
+			remoteIterators.forEach(this::close);
 			connectionHandler.close();
 		}
 		closed.accept(this);
@@ -173,7 +204,75 @@ public abstract class AbstractRemoteEntityConnection extends UnicastRemoteObject
 		return closed.observer();
 	}
 
+	final RemoteEntityResultIterator remoteIterator(EntityResultIterator iterator) throws RemoteException {
+		return new DefaultRemoteEntityResultIterator(connectionPort, clientSocketFactory, serverSocketFactory, iterator);
+	}
+
+	final void cleanupIterators() {
+		remoteIterators.stream()
+						.filter(DefaultRemoteEntityResultIterator::timedOut)
+						.forEach(this::close);
+	}
+
+	private void close(RemoteEntityResultIterator iterator) {
+		try {
+			LOG.debug("Closing iterator for {}", user());
+			iterator.close();
+		}
+		catch (Exception e) {
+			LOG.error("Failed to close iterator for {}: {}", user(), e.getMessage(), e);
+		}
+	}
+
 	static int requestsPerSecond() {
 		return LocalConnectionHandler.REQUEST_COUNTER.requestsPerSecond();
+	}
+
+	private final class DefaultRemoteEntityResultIterator extends UnicastRemoteObject implements RemoteEntityResultIterator {
+
+		@Serial
+		private static final long serialVersionUID = 1;
+
+		private final EntityResultIterator iterator;
+
+		private final long timeout = ITERATOR_TIMEOUT.getOrThrow();
+
+		private long lastAccessTime = currentTimeMillis();
+
+		private DefaultRemoteEntityResultIterator(int port, RMIClientSocketFactory clientSocketFactory,
+																							RMIServerSocketFactory serverSocketFactory,
+																							EntityResultIterator iterator) throws RemoteException {
+			super(port, clientSocketFactory, serverSocketFactory);
+			this.iterator = iterator;
+			remoteIterators.add(this);
+		}
+
+		@Override
+		public boolean hasNext() throws RemoteException {
+			lastAccessTime = currentTimeMillis();
+
+			return iterator.hasNext();
+		}
+
+		@Override
+		public Entity next() throws RemoteException {
+			lastAccessTime = currentTimeMillis();
+
+			return iterator.next();
+		}
+
+		@Override
+		public void close() throws RemoteException {
+			try {
+				iterator.close();
+			}
+			catch (Exception ignored) {/*ignored*/}
+			unexportObject(this, true);
+			remoteIterators.remove(this);
+		}
+
+		boolean timedOut() {
+			return currentTimeMillis() - lastAccessTime > timeout;
+		}
 	}
 }